Added setters for log enablement, all-log enablement, and log level.
To make this make sense, also had to add code that checks whether
logs are enabled at call time rather than checking the value of
`IsEnabled` which is set at instantiation time. The value of IsEnabled
is updated whenever an internal method that uses it checks for
enablement.

Added a little bit of documentation to existing and new functions.